Modified type of fuzz target to usize

+++ fuzz/fuzz_targets/bytes_exterior.rs
@@ -5,9 +5,10 @@ use libfuzzer_sys::fuzz_target;
 use bytes::{BytesMut, BufMut};
 
 
-fuzz_target!(|input: &[u8]| {
+fuzz_target!(|input: &[usize]| {
 
     let mut buf = BytesMut::with_capacity(input);
     buf.put(&b"hello world"[..]);
     buf.put_u16(1234);
+
 });
